requirements.txt
asgiref==3.2.3
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
fake-factory==9999.9.9
Faker==4.0.0
Markdown==3.2
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
XlsxWriter==1.2.7
